+++ docs/app/[lng]/[...not-found]/page.tsx
@@ -43,7 +43,7 @@ const NotFoundPage = (props: PageProps) => {
 	const lng = params?.lng ?? LocaleConfig.default
 
 	return (
-		<Grid bgcolor="grey.100">
+		<Grid bgcolor="background.default">
 			<Container maxWidth="sm" sx={{ display: 'flex', alignItems: 'center', height: '100dvh' }}>
 				<Paper variant="outlined" sx={{ width: 1, borderRadius: 2 }}>
 					<Grid container flexDirection="column" alignItems="center" spacing={2} py={4} px={{ xs: 1, md: 2 }}>