+++ docs/types/components/molecules/list.d.ts
@@ -2,11 +2,12 @@
 import type { ReactNode } from 'react'
 
 export type ListItemProp = {
-	title: string
+	title?: string
 	link?: string
 	icon?: ReactNode
 	iconColor?: string
 	children?: ListItem[]
+	onClick?: () => void
 }
 
 export type ListProps = {