+++ docs/types/components/organisms/appbar.d.ts
@@ -1,1 +1,1 @@
-export type AppbarProps = { lng?: string }
+export type AppbarProps = { lng: string }